requirements.txt
streamlit
streamlit-webrtc
opencv-python-headless
ultralytics
Pillow
av
numpy